+++ include/libcopp/coroutine/generator_promise.h
@@ -50,6 +50,12 @@ class LIBCOPP_COPP_API_HEAD_ONLY generator_future;
 template <class TPROMISE, bool RETURN_VOID, generator_vtable_type VTABLE_TYPE>
 class LIBCOPP_COPP_API_HEAD_ONLY generator_awaitable;
 
+template <class TVALUE, class TERROR_TRANSFORM = promise_error_transform<TVALUE>>
+using generator_lightweight_future = generator_future<TVALUE, TERROR_TRANSFORM, generator_vtable_type::kLightWeight>;
+
+template <class TVALUE, class TERROR_TRANSFORM = promise_error_transform<TVALUE>>
+using generator_channel_future = generator_future<TVALUE, TERROR_TRANSFORM, generator_vtable_type::kNone>;
+
 template <class TVALUE>
 class LIBCOPP_COPP_API_HEAD_ONLY generator_context_base {
  public:
@@ -724,6 +730,21 @@ class LIBCOPP_COPP_API_HEAD_ONLY some_delegate<generator_future<TVALUE, TERROR_T
 
   using base_type::run;
 };
+
+template <class TVALUE, class TERROR_TRANSFORM = promise_error_transform<TVALUE>>
+using generator_channel_receiver = generator_channel_future<TVALUE, TERROR_TRANSFORM>;
+
+template <class TVALUE, class TERROR_TRANSFORM = promise_error_transform<TVALUE>>
+using generator_channel_sender = typename generator_channel_receiver<TVALUE, TERROR_TRANSFORM>::context_pointer_type;
+
+template <class TVALUE, class TERROR_TRANSFORM = promise_error_transform<TVALUE>>
+inline std::pair<generator_channel_receiver<TVALUE, TERROR_TRANSFORM>,
+                 generator_channel_sender<TVALUE, TERROR_TRANSFORM>>
+make_channel() {
+  generator_channel_receiver<TVALUE, TERROR_TRANSFORM> future;
+  generator_channel_sender<TVALUE, TERROR_TRANSFORM> context = future.get_context();
+  return std::make_pair(std::move(future), std::move(context));
+}
 
 LIBCOPP_COPP_NAMESPACE_END